+++ src/backend/message/reaction_message.rs
@@ -1,5 +1,6 @@
 use std::cell::RefCell;
 
+use gdk::gdk_pixbuf::Pixbuf;
 use gio::subclass::prelude::ObjectSubclassIsExt;
 use glib::Object;
 use gtk::{gio, glib};
@@ -8,7 +9,9 @@ use presage::prelude::*;
 
 use crate::backend::{timeline::TimelineItem, Channel, Contact};
 
-use super::{Manager, Message};
+use super::{Manager, Message, MessageExt};
+
+use gettextrs::gettext;
 
 gtk::glib::wrapper! {
     pub struct ReactionMessage(ObjectSubclass<imp::ReactionMessage>) @extends Message, TimelineItem;
@@ -50,6 +53,39 @@ impl ReactionMessage {
 
     pub fn target_uuid(&self) -> Uuid {
         Uuid::parse_str(self.reaction().target_author_uuid()).expect("`Reaction` Uuid to be valid")
+    }
+
+    pub fn send_notification(&self) {
+        let sender = self.sender();
+        let channel = self.channel();
+        if sender.is_self() || self.target_uuid() != self.manager().uuid() {
+            // Skip notifications for reaction messages sent from self.
+            return;
+        }
+        let notification_title;
+        let notification_body;
+        if channel.group_context().is_some() {
+            notification_title = channel.title();
+            // Translators: When receiving a reaction message in a group, this will be the text to format the notification. Do not translate the text in {}.
+            notification_body = gettext("{sender} reacted {emoji} to a message.")
+                .replace("{sender}", &sender.title())
+                .replace("{emoji}", &self.emoji());
+        } else {
+            notification_title = sender.title();
+            // Translators: When receiving a reaction message in a 1-to-1 channel (the body, the sender will be in the notification title and is not included in the translation), this will be the text to format the notification. Do not translate the text in {}.
+            notification_body = gettext("Reacted {emoji} to a message.")
+                .replace("{sender}", &sender.title())
+                .replace("{emoji}", &self.emoji());
+        }
+        let notification = gio::Notification::new(&notification_title);
+        notification.set_body(Some(&notification_body));
+        let icon = Pixbuf::from_resource("/icon.png").expect("Flare to have an application icon");
+        notification.set_icon(&icon);
+
+        let manager = self.manager();
+        crate::gspawn!(async move {
+            manager.send_notification(&notification).await;
+        });
     }
 }